ref: move dpe error definition into separate file
