+++ dpe/src/commands/derive_context.rs
@@ -5,7 +5,8 @@ use crate::{
     context::{ActiveContextArgs, Context, ContextHandle, ContextState, ContextType},
     dpe_instance::{DpeEnv, DpeInstance},
     mutresp, okref,
-    response::{DeriveContextExportedCdiResp, DeriveContextResp, DpeErrorCode, InternalErrorCode},
+    error::{DpeErrorCode, InternalErrorCode},
+    response::{DeriveContextExportedCdiResp, DeriveContextResp},
     tci::TciMeasurement,
     x509::{create_exported_dpe_cert, CreateDpeCertArgs, CreateDpeCertResult},
     DpeFlags, State,
