+++ dpe/src/commands/get_certificate_chain.rs
@@ -3,7 +3,8 @@ use super::CommandExecution;
 use crate::{
     dpe_instance::{DpeEnv, DpeInstance},
     mutresp,
-    response::{DpeErrorCode, GetCertificateChainResp},
+    error::DpeErrorCode,
+    response::GetCertificateChainResp,
 };
 #[cfg(feature = "cfi")]
 use caliptra_cfi_derive::cfi_impl_fn;
